backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptograpyh]
passlib[bcrypt] == 1.7.4
sqlalchemy
python-dotenv
pydantic[email]
supabase
